refactor: migrate keyboard driver to Nitrogen PS/2 implementation
Remove custom scancode-to-ASCII conversion, input buffering, and PS/2 probe/initialization. Replace with Nitrogen's PS/2 keyboard driver for unified device handling and maintainability. Update interrupt handler and init sequence accordingly.
