+++ fullerene-kernel/src/init.rs
@@ -92,6 +92,11 @@ pub fn init_common(physical_memory_offset: x86_64::VirtAddr) {
                 }
             }
         }),
+        petroleum::init_step!("PS2 Keyboard", || {
+            nitrogen::ps2::keyboard::init_keyboard();
+            petroleum::serial::serial_log(format_args!("PS/2 keyboard initialised\n"));
+            Ok(())
+        }),
         petroleum::init_step!("process", || {
             let heap_start =
                 unsafe { core::ptr::addr_of_mut!(crate::heap::BOOT_HEAP_BUFFER) as usize };
