+++ nitrogen/src/ps2/mod.rs
@@ -3,4 +3,5 @@
 //! This module provides drivers for PS/2 peripherals (mouse, keyboard, etc.)
 //! using the underlying port I/O primitives available in Nitrogen.
 
+pub mod keyboard;
 pub mod mouse;
